switch authStatusReducer to ts
