+++ src/components/cities-map/cities-map.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {connect} from 'react-redux';
 import MapComponent from '../map-component/map-component';
 import {getCityOffers} from '../../store/reducers/offersReducer/selectors';
-import {getActiveOffer} from '../../store/reducers/activeOfferReducer/selectors';
+import {getActiveOffer} from '../../store/reducers/active-offer-reducer/selectors';
 import {getCurrentCity} from '../../store/reducers/currentCityReducer/selectors';
 import {IPlace, ICity, IStore} from '../../types/types';
