+++ src/components/offer-details/offer-details.tsx
@@ -9,8 +9,8 @@ import ReviewForm from '../review-form/review-form';
 import {getOfferByParamsId} from '../../store/reducers/offersReducer/selectors';
 import {getOffersNearby} from '../../store/reducers/offers-nearby-reducer/selectors';
 import {getComments} from '../../store/reducers/commentsReducer/selectors';
-import {getAuthStatus} from '../../store/reducers/authStatusReducer/selectors';
-import {AUTH_STATUS} from '../../store/reducers/authStatusReducer/authStatusReducer';
+import {getAuthStatus} from '../../store/reducers/auth-status-reducer/selectors';
+import {AUTH_STATUS} from '../../store/reducers/auth-status-reducer/auth-status-reducer';
 import OffersOperation from '../../store/operations/offers-operation/offers-operation';
 import CommentsOperation from '../../store/operations/comments-operation/comments-operation';
 import {IPlace, IReview, IStore} from '../../types/types';
