+++ src/store/reducers/active-offer-reducer/active-offer-reducer.ts
