+++ src/store/reducers/active-offer-reducer/selectors.ts
@@ -1,0 +1,6 @@
+import NameSpace from '../name-space';
+import {IStore} from '../../../types/types';
+
+export const getActiveOffer = (state: IStore) => {
+  return state[NameSpace.ACTIVE_OFFER];
+};
