--- src/store/reducers/activeOfferReducer/selectors.js
@@ -1,5 +1,0 @@
-import NameSpace from '../name-space';
-
-export const getActiveOffer = (state) => {
-  return state[NameSpace.ACTIVE_OFFER];
-};
