+++ src/store/reducers/auth-status-reducer/selectors.ts
@@ -1,0 +1,6 @@
+import NameSpace from '../name-space';
+import {IStore} from '../../../types/types';
+
+export const getAuthStatus = (state: IStore) => {
+  return state[NameSpace.AUTH_STATUS];
+};
