--- src/store/reducers/authStatusReducer/selectors.js
@@ -1,5 +1,0 @@
-import NameSpace from '../name-space';
-
-export const getAuthStatus = (state) => {
-  return state[NameSpace.AUTH_STATUS];
-};
