+++ src/store/reducers/reducer.js
@@ -1,10 +1,10 @@
 import {combineReducers} from 'redux';
 import currentCityReducer from './currentCityReducer/currentCityReducer';
-import activeOfferReducer from './activeOfferReducer/active-offer-reducer';
+import activeOfferReducer from './active-offer-reducer/active-offer-reducer';
 import sortingOptionReducer from './sortingOptionReducer/sortingOptionReducer';
 import citiesReducer from './citiesReducer/citiesReducer';
 import offersReducer from './offersReducer/offersReducer';
-import authStatusReducer from './authStatusReducer/authStatusReducer';
+import authStatusReducer from './auth-status-reducer/auth-status-reducer';
 import userReducer from './userReducer/userReducer';
 import offersNearbyReducer from './offers-nearby-reducer/offers-nearby-reducer';
 import sendingCommentReducer from './sending-comment-reducer/sending-comment-reducer';
